Undo commit 0837c1049a33e772211e8ce12260203565341296
Had previously modified core.js to use CommonJS export, so
we could do `var d3 = require("d3.js");` however this is
a bad idea, because the d3 submodules expect the variable
d3 to be in the global namespace. We therefore revert to the
original method of just adding d3 to the global namespace.